Internals

Internal Interface

Documentation for JuDoc.jl's internal interface.

File processing and compilation

JuDoc.jd_setup — Function.

jd_setup(; clear)

Sets up the collection of watched files by doing an initial scan of the input directory. It also sets the paths variables and prepares the output directory.

Keyword argument

clear=false: whether to remove any existing output directory

See also serve.

source

JuDoc.jd_fullpass — Function.

jd_fullpass(watched_files; clear, verb, prerender, isoptim)

A single full pass of judoc looking at all watched files and processing them as appropriate.

Keyword arguments

clear=false: whether to remove any existing output directory

verb=false: whether to display messages

prerender=false: whether to prerender katex and code blocks

See also jd_loop, serve and publish.

source

JuDoc.jd_loop — Function.

jd_loop(cycle_counter, #temp#, watched_files; clear, verb)

This is the function that is continuously run, checks if files have been modified and if so, processes them. Every 30 cycles, it checks whether any file was added or deleted and consequently updates the watched_files.

Keyword arguments

clear=false: whether to remove any existing output directory

verb=false: whether to display messages

source

JuDoc.process_file — Function.

process_file(case, fpair, args; kwargs...)

See process_file_err.

source

JuDoc.process_file_err — Function.

process_file_err(case, fpair)
process_file_err(case, fpair, head)
process_file_err(case, fpair, head, pg_foot)
process_file_err(case, fpair, head, pg_foot, foot)
process_file_err(case, fpair, head, pg_foot, foot, t; clear, prerender, isoptim)

Considers a source file which, depending on case could be a html file or a file in judoc markdown etc, located in a place described by fpair, processes it by converting it and adding appropriate header and footer and writes it to the appropriate place. It can throw an error which will be caught in process_file(args...).

source

Conversion

JuDoc.write_page — Function.

write_page(root, file, head, pg_foot, foot; prerender, isoptim)

Take a path to an input markdown file (via root and file), then construct the appropriate HTML page (inserting head, pg_foot and foot) and finally write it at the appropriate place.

source

JuDoc.convert_md — Function.

convert_md(mds)
convert_md(mds, pre_lxdefs; isrecursive, isconfig, has_mddefs)

Convert a judoc markdown file read as mds into a judoc html string. Returns the html string as well as a dictionary of page variables.

Arguments

mds: the markdown string to process

pre_lxdefs: a vector of LxDef that are already available.

Keyword arguments

isrecursive=false: a bool indicating whether the call is the parent call or a child call

isconfig=false: a bool indicating whether the file to convert is the configuration file

has_mddefs=true: a bool indicating whether to look for definitions of page variables

source

JuDoc.form_inter_md — Function.

form_inter_md(mds, blocks, lxdefs)

Form an intermediate MD file where special blocks are replaced by a marker (JD_INSERT) indicating that a piece will need to be plugged in there later.

Arguments

mds: the (sub)string to convert

blocks: vector of blocks

lxdefs: existing latex definitions prior to the math block

source

JuDoc.md2html — Function.

md2html(ss)
md2html(ss, stripp)

Convenience function to call the base markdown to html converter on "simple" strings (i.e. strings that don't need to be further considered and don't contain anything else than markdown tokens). The boolean stripp indicates whether to remove the inserted <p> and </p> by the base markdown processor, this is relevant for things that are parsed within latex commands etc.

source

JuDoc.convert_inter_html — Function.

convert_inter_html(ihtml, blocks, lxcontext)

Take a partial markdown string with the JD_INSERT marker and plug in the appropriately processed block.

Arguments

ihtml: the intermediary html string (with JD_INSERT)

blocks: vector of blocks

lxcontext: latex context

source

JuDoc.convert_block — Function.

convert_block(β, lxcontext)

Helper function for convert_inter_html that processes an extracted block given a latex context lxc and returns the processed html that needs to be plugged in the final html.

source

JuDoc.convert_html — Function.

convert_html(hs, allvars)
convert_html(hs, allvars, fpath; isoptim)

Convert a judoc html string into a html string (i.e. replace {{ ... }} blocks).

source

JuDoc.convert_hblock — Function.

convert_hblock(β, allvars)
convert_hblock(β, allvars, fpath)

Helper function to process an individual block when the block is a HCond such as {{ if showauthor }} {{ fill author }} {{ end }}.

source

convert_hblock(β, allvars)
convert_hblock(β, allvars, fpath)

Helper function to process an individual block when the block is a HIsDef such as {{ ifdef author }} {{ fill author }} {{ end }}. Which checks if a variable exists and if it does, applies something.

source

convert_hblock(β, allvars)
convert_hblock(β, allvars, fpath)

Helper function to process an individual block when the block is a HIsPage such as {{ ispage path/to/page}} ... {{end}}. Which checks if the current page is a given one and applies something if that's the case (useful to handle different layouts on different pages).

source

convert_hblock(β, allvars)
convert_hblock(β, allvars, #temp#)

Helper function to process an individual block when it's a HFun such as {{ fill author }}. Dev Note: fpath is (currently) unused but is passed to all convert_hblock functions. See convert_html.

source

JuDoc.build_page — Function.

build_page(head, content, pg_foot, foot)

Convenience function to assemble the html out of its parts.

source

Pre-rendering

JuDoc.js_prerender_katex — Function.

js_prerender_katex(hs)

Takes a html string that may contain inline katex blocks \(...\) or display katex blocks \[ ... \] and use node and katex to pre-render them to HTML.

source

JuDoc.js_prerender_highlight — Function.

js_prerender_highlight(hs)

Takes a html string that may contain <pre><code ... </code></pre> blocks and use node and highlight.js to pre-render them to HTML.

source

Parsing

JuDoc.AbstractBlock — Type.

abstract type AbstractBlock

This is the supertype defining a section of the string into consideration which can be considered as a "block". For instance, abc { ... } def contains one braces block corresponding to the substring { ... }. All subtypes of AbstractBlock must have a ss field corresponding to the substring associated to the block. See also Token, OCBlock.

source

JuDoc.Token — Type.

struct Token <: JuDoc.AbstractBlock

A token τ::Token denotes a part of the source string indicating a region that may need further processing. It is identified by a symbol τ.name (e.g.: :MATH_ALIGN_OPEN). Tokens are typically used in this code to identify delimiters of environments. For instance, abc $ ... $ def contains two tokens :MATH_A associated with the $ sign. Together they delimit here an inline math expression.

source

JuDoc.OCBlock — Type.

struct OCBlock <: JuDoc.AbstractBlock

Open-Close block, blocks that are defined by an opening Token and a closing Token, they may be nested. For instance braces block are formed of an opening { and a closing } and could be nested.

source

JuDoc.TokenFinder — Type.

TokenFinder

Convenience type to define tokens. The Tuple comes from the output of functions such as isexactly.

source

JuDoc.MD_TOKENS — Constant.

MD_TOKENS

Dictionary of tokens for Markdown. Note that for each, there may be several possibilities to consider in which case the order is important: the first case that works will be taken.

source

JuDoc.MD_1C_TOKENS — Constant.

MD1CTOKENS

Dictionary of single-char tokens for Markdown. Note that these characters are exclusive, they cannot appear again in a larger token.

source

JuDoc.LxDef — Type.

mutable struct LxDef

Structure to keep track of the definition of a latex command declared via a ewcommand{ ame}[narg]{def}.

NOTE: mutable so that we can modify the from element to mark it as zero when the command has been defined in the context of what we're currently parsing.

source

JuDoc.LxCom — Type.

struct LxCom <: JuDoc.AbstractBlock

A LxCom has a similar content as a Block, with the addition of the definition and a vector of brace blocks.

source

JuDoc.HTML_TOKENS — Constant.

HTML_TOKENS

Dictionary of tokens for HTML. Note that for each, there may be several possibilities to consider in which case the order is important: the first case that works will be taken.

source

JuDoc.HTML_OCB — Constant.

HTML_OCB

List of HTML Open-Close blocks.

source

JuDoc.isexactly — Function.

isexactly(refstring)
isexactly(refstring, follow)
isexactly(refstring, follow, isfollowed)

Forward lookup checking if a sequence of characters matches refstring and is followed (or not followed if isfollowed==false) by a character out of a list of characters (follow). It returns

a number of steps indicating the number of characters to check,

whether there is an offset or not (if it is required to check a following character or not),

a function that can be applied on a sequence of character.

Example

julia> (s, b, f) = isexactly("aabbc", "d", isfollowed=false);
julia> f("aabbcd")
false
julia> f("aabbce")
true
julia> s
5

source

JuDoc.find_tokens — Function.

find_tokens(str, tokens_dict, stokens_dict)

Go through a text left to right, one (valid) char at the time and keep track of sequences of chars that match specific tokens. The list of tokens found is returned.

Arguments

str: the initial text

tokens_dict: dictionary of possible tokens (multiple characters long)

stokens_dict: dictionaro of possible tokens (single character)

source

JuDoc.find_all_ocblocks — Function.

find_all_ocblocks(tokens, ocblist; inmath)

Convenience function to find all ocblocks e.g. such as MD_OCBLOCKS. Returns a vector of vectors of ocblocks.

source

JuDoc.find_md_lxdefs — Function.

find_md_lxdefs(tokens, blocks)

Find \newcommand elements and try to parse what follows to form a proper Latex command. Return a list of such elements.

The format is: \newcommand{NAMING}[NARG]{DEFINING} where [NARG] is optional (see LX_NARG_PAT).

source

JuDoc.find_md_lxcoms — Function.

find_md_lxcoms(tokens, lxdefs, braces)
find_md_lxcoms(tokens, lxdefs, braces, offset; inmath)

Find \command{arg1}{arg2}... outside of xblocks and lxdefs.

source

JuDoc.qualify_html_hblocks — Function.

qualify_html_hblocks(blocks)

Given {{ ... }} blocks, identify what kind of blocks they are and return a vector of qualified blocks of type AbstractBlock.

source

JuDoc.find_html_cblocks — Function.

find_html_cblocks(qblocks)

Given qualified blocks HIf, HElse etc, construct a vector of the conditional blocks which contain the list of conditions etc. No nesting is allowed at the moment.

source

JuDoc.find_html_cdblocks — Function.

find_html_cdblocks(qblocks)

Given qualified blocks HIsDef or HIsNotDef build conditional page blocks.

source

JuDoc.find_html_cpblocks — Function.

find_html_cpblocks(qblocks)

Given qualified blocks HIsPage or HIsNotPage build conditional page blocks.

source