Public

Public Interface

Documentation for JuDoc.jl's exported functions

JuDocTemplates.newsite — Function.

newsite(topdir; template="basic", cd=true)

Generate a new folder (an error is thrown if it already exists) that contains the skeleton of a website that can be processed by JuDoc. The user can specify a template out of the list of available templates.

template="basic" specify the name of the desired template,

cd=true specify whether to change the current directory to the newly created folder or not.

verbose=true specify whether to display information or not.

Example

newsite("MyNewWebsite", template="pure-sm")

JuDoc.serve — Function.

serve(; clear, verb, port, single, prerender, nomess)

Runs JuDoc in the current directory.

Keyword arguments:

clear=false: whether to remove any existing output directory

verb=false: whether to display messages

port=8000: the port to use for the local server (should pick a number between 8000 and 9000)

single=false: whether to run a single pass or run continuously

prerender=false: whether to pre-render javascript (KaTeX and highlight.js)

nomess=false: suppresses all messages (internal use).

source

JuDoc.optimize — Function.

optimize(; prerender, minify, sig)

Does a full pass followed by a pre-rendering and minification step.

prerender=true: whether to pre-render katex and highlight.js (requires node.js)

minify=true: whether to minify output (requires python3 and css_html_js_minify)

sig=false: whether to return an integer indicating success (see publish)

Note: if the prerendering is set to true, the minification will take longer as the HTML files will be larger (especially if you have lots of maths on pages).

source

JuDoc.publish — Function.

publish(; prerender, minify, nopass)

This is a simple wrapper doing a git commit and git push without much fanciness. It assumes the current directory is a git folder. This will work in most simple scenarios (e.g. there's only one person updating the website). In other scenarios you should probably do this manually.

Keyword arguments

prerender=true: prerender javascript before pushing see optimize

minify=true: minify output before pushing see optimize

nopass=false: set this to true if you have already run optimize manually.

source

JuDoc.cleanpull — Function.

cleanpull()

Cleanpull allows you to pull from your remote git repository after having removed the local output directory. This will help avoid merge clashes.

source